+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Element injection into options nested in repeatable sets
- - `$input-border-color` CSS override
+- `$input-border-color` CSS override
 
 ### Changed
 
 - `pointCadastralParcel` to only show results grid if address has `features`
+- `$danger` CSS property to meet accessibility requirements
 
 ## [8.11.2] - 2025-10-13