+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Element injection into options nested in repeatable sets
 - `$input-border-color` CSS override
+- default restricted file types
 
 ### Changed